SPS-Master/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.6
django-bootstrap5==22.2
gcloud==0.18.3
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
oauth2client==4.1.3
protobuf==4.22.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.14.1
